lightning: move next_step logic to fabric
