+++ src/dvclive/fabric.py
@@ -1,3 +1,4 @@
+import inspect
 from typing import TYPE_CHECKING, Any, Dict, Mapping, Optional, Union
 
 try:
@@ -24,6 +25,24 @@ from dvclive.utils import standardize_metric_name
 
 if TYPE_CHECKING:
     from dvclive import Live
+
+
+def _should_call_next_step():
+    """
+    Find out if pytorch_lightning is calling `log_metrics` from the functions
+    where we actually want to call `next_step`.
+    For example, prevents calling next_step when external callbacks call
+    `log_metrics` or during the multiple `update_eval_step_metrics`.
+    """
+    return any(
+        frame.function
+        in (
+            "update_train_step_metrics",
+            "update_train_epoch_metrics",
+            "log_eval_end_metrics",
+        )
+        for frame in inspect.stack()
+    )
 
 
 class DVCLiveLogger(Logger):
@@ -90,6 +109,14 @@ class DVCLiveLogger(Logger):
                     f"\n you tried to log {val} which is currently not supported."
                     "Try a scalar/tensor."
                 )
+
+        if _should_call_next_step():
+            if step == self.experiment._latest_studio_step:  # noqa: SLF001
+                # We are in log_eval_end_metrics but there has been already
+                # a studio request sent with `step`.
+                # We decrease the number to bypass `live.studio._get_unsent_datapoints`
+                self.experiment._latest_studio_step -= 1  # noqa: SLF001
+            self.experiment.next_step()
 
     @rank_zero_only
     def log_hyperparams(self, params: Union[Dict[str, Any]]) -> None:
